feat: add name update API and support for new salary calculation unit

+++ src/main/java/com/api/alba/dto/owner/OwnerMonthlyCalendarItemResponse.java
@@ -14,6 +14,6 @@ public class OwnerMonthlyCalendarItemResponse {
     private String profileColor;
     private String profileInitial;
     private LocalDate workDate;
-    private Integer workedHours;
+    private BigDecimal workedHours;
     private BigDecimal finalWage;
 }
